Fix typo in TrashCanIcon
