+++ src/components/icons/TrashCanIcon.tsx
