+++ src/index.ts
@@ -84,7 +84,7 @@ export { default as StatusOkIcon } from './components/icons/StatusOkIcon'
 export { default as SuitcaseIcon } from './components/icons/SuitcaseIcon'
 export { default as TableIcon } from './components/icons/TableIcon'
 export { default as TagIcon } from './components/icons/TagIcon'
-export { default as TashCanIcon } from './components/icons/TashCanIcon'
+export { default as TrashCanIcon } from './components/icons/TrashCanIcon'
 export { default as TwitterIcon } from './components/icons/TwitterIcon'
 export { default as UpdatesIcon } from './components/icons/UpdatesIcon'
 export { default as VolumesIcon } from './components/icons/VolumesIcon'
